Cycles: Replace thin-film basis function approximation with accurate LUTs
Previously, we used precomputed Gaussian fits to the XYZ CMFs, performed the
spectral integration in that space, and then converted the result to the RGB
working space.

That worked because we're only supporting dielectric base layers for the thin
film code, so the inputs to the spectral integration (reflectivity and phase)
are both constant w.r.t. wavelength.

However, this will no longer work for conductive base layers. We could handle
reflectivity by converting to XYZ, but that won't work for phase since its
effect on the output is nonlinear.

Therefore, it's time to do this properly by performing the spectral integration
directly in the RGB primaries. To do this, we need to:
- Compute the RGB CMFs from the XYZ CMFs and XYZ-to-RGB matrix
- Resample the RGB CMFs to be parametrized by frequency instead of wavelength
- Compute the FFT of the CMFs
- Store it as a LUT to be used by the kernel code

None of that is particularly hard, but we need a FFT implementation for it.
I've kept it simple for now since it's just a single N=512 FFT, so pulling in
e.g. FFTW for it isn't worth it.

Output is mostly the same, just slightly different because we're no longer
using the Gaussian approximation.