Build: Enable Embree debug compilation with SYCL support
Previously, Embree debug compilation with SYCL support was
not working properly, so it was disabled. Now, after the upgrade
of the DPC++ compiler, all previously noted problems seem to be
addressed, so this is enabled back for a better debugging experience.

Pull Request: https://projects.blender.org/blender/blender/pulls/145400