Deps: Upgrade IGC, Ocloc to new versions with support for upcoming HW
These changes are upgrading IGC (Intel GPU Compiler) to the version
v2.16.0 and Ocloc (OpenCL Offline Compiler) to the version
25.31.34666.3 (correct version string is 25.31.34666+3).
This is required for the proper support of the upcoming Intel GPUs,
in order for Blender users to be able to use Blender 5.0 with it
in the future.

Overall, these changes keep the performance on the same level,
with minor (1%) improvements in some scenes, with the only
side-effect that we need to raise the minimal driver
on Linux from 25.31.31740 to 25.31.34666, which has a minimal
support OS version as Ubuntu 24.04 LTS, while the previous driver
minimal version was available on Ubuntu 22.04 LTS and older,
which makes this commit a support breaking commit - users on
22.04 LTS should upgrade to 24.04 LTS if they plan to use
Blender 5.0 with Intel GPUs.

There are no expected changes for Windows execution, this is
purely a Linux commit.

Pull Request: https://projects.blender.org/blender/blender/pulls/145539