Deps: Update OpenXR SDK to 1.1.53 and enable OpenXR on macOS
This PR updates the OpenXR SDK library from version 1.0.22 to 1.1.53 and
enables it on macOS (prior to this OpenXR was only built and supported
on Windows and Linux), this follows up on the recent development of an
OpenXR macOS Metal graphics binding (#142626).

Tested the library to build and work correctly on macOS arm64, using the
said graphics binding and the Meta XR Simulator.

Pull Request: https://projects.blender.org/blender/blender/pulls/148217